cty: Fix deep marks functions with new transformer
The UnmarkDeep, UnmarkDeepWithPaths, and MarkWithPaths functions could
previously panic with nested complex values. This was due to the
Transform callback function only being called as part of a postorder
traversal, when it is not permitted to traverse a marked value.

This commit introduces a new TransformWithTransformer function,
requiring an implementation of a new Transformer interface. Using this
interface, callers can implement either preorder or postorder
traversals.

In turn, this allows us to write deep transformation for marks which
successfully cope with all nested structures.

The commit includes significantly more tests for these functions, which
should now cover all complex value types.