cty: Add WithMarksDeep
This is to WithMarks as UnmarkDeep is to Unmark. When applying marks to
nested values, it can be useful to recursively apply marks to the entire
tree. Consider a "tainted" mark applied to a complex object value:
carrying this through to every leaf value ensures that no "tainted"
values leak as the object is traversed.