function: Add support for marked elements to join
Previously the join function would correctly handle marks on the
separator or the given lists of strings, but panic if any of the lists
contained strings which were marked. This commit adds support for this
last case.

Calling `join(a, [b, c, d])` will return a string value marked with the
union of all marks for the separator, list, and list elements.