convert: Fix MismatchMessage for object attributes
When determining which object attributes cause a type mismatch, we must
disregard equal attribute types. Previously we would attempt to find a
conversion between them, deciding that the attribute type was a mismatch
if no conversion exists. However, this is invalid if the two types are
equal, because GetConversion returns nil for equal primitive types.

Because of non-deterministic map iteration order, the test case added in
this commit failed approximately 2/3 of the time before the change. It
now passes consistently.