parseint: return a NumberIntVal
parseint was returning a cty.NumberVal - a float - which caused some odd
behavior when the result, which users expected to be an integer, was used
in other math functions. See the upstream terraform issue
https://github.com/hashicorp/terraform/issues/26365 for some background
context.

The test case added was the example large number that was returning an
unexpected result from parseint.