cty: Allow list of vals with partly-dynamic types
The ListVal constructor function previously required that all elements
had exactly the same type. In some cases involving partly-dynamic
element types, this was overly strict. This commit relaxes the
constraints which caused these cases to panic.

If we attempted to create a list of objects, some of which contained
dynamic pseudo-type fields, the constructor would panic. For example,
consider this (HCL) example:

list = [
  {
    ids = [1, 2],
  },
  {
    ids = [for obj in some_dynamic_list: obj.id],
  },
]

This panic happened because the two element types were different:
namely, Object({ ids: string }) and Object({ ids: dynamic }). However,
these types are unifiable, and the convert package calls ListVal after
verifying this.

Instead of strictly comparing all element types, we now only panic if
the list contains two or more elements with fully concrete types which
are inequal. Element types which are dynamic or partly dynamic are
permitted.