cty/function: Preserve marks when short-circuiting for unknown
Our functions mechanism has some default behaviors for both unknown values
and marked values in order to get the most common behaviors automatically
and avoid re-implementing them separately in each function.

However, we were not handling the intersection of those two cases properly.
If we aren't going to call the real function implementation at all, as
is the case if there's at least one unknown value, we must behave as if
the function can't handle marked values on any of its arguments because
that's the fullest set of marks the real function could potentially have
generated, and we err on the side of conservatively retaining marks that
we're not sure about yet.

This is covered indirectly by a test for the Negate stdlib function for
now, because that's what originally drew attention to this edge case.