convert: Fix for null objects with optional attrs
Conversions targeting types which are or include objects with optional
attributes always simplify those types, removing the optional attributes
from the type, and inserting null values for any that are missing.
Objects with optional attributes are not intended to be used to
instantiate values.

One accidental exception to this was the null value of dynamic type.
These values are passed through by the getConversion wrapper, assigning
the desired type. This conflicts with the type simplification rules for
non-null values, and can result in a panic when a collection type
contains null values.

This commit fixes this at the wrapper level by performing the same type
simplification when converting null into an object type.