special case unifying tuples as lists
List data is often contained within tuple types, either do to language
constructs (hcl tuple literals) or function return types. When presented
with only lists and tuples for unification, first check if the tuples
themselves unify as lists. If they do, run the unification again with
only the list types, and wrap the conversions as necessary.