convert: Recursively concretize unknown/null
When converting unknown and null values to a given type, we want to
avoid constructing instances of types which include optional object
attributes, even deeply nested. To do this we add a new method,
cty.Type.WithoutOptionalAttributesDeep, which recursively processes its
receiver to return a type which has no instances of objects with
optional attributes.