convert: Don't panic when collection element type unification fails
Despite our best efforts to correctly unify element types during conversion,
sometimes some inconsistent types slip through anyway and make the
conversion panic.

Although in the long run it would be better to fix these problems upstream,
it's the responsibility of the caller of cty to not try to perform any
invalid operations, and so here we make the convert package pre-verify
that it's found a sensible result before trying to construct it, and
thus it can return a proper error if not rather than panicking.