function/stdlib: ValuesFunc precise handling of marks
Previously ValuesFunc was just using the safe default behavior where if
any argument contains any marked values at all then the result is
automatically given the same marks.

That's overly conservative though, because it means that an unmarked
map with a single marked element value will become a wholly-marked list
result, rather than an unmarked list with one marked element value.

Now we'll opt in to custom mark handling within the function itself, and
handle shallowly the marking of the overall collection -- propagating it
to the result -- while letting the values inside retain their individual
markings as needed.

We were previously lacking unit tests for the values function, I think
because it's a pretty old one from when cty was new, so this also includes
additional test cases that are not directly related to marked values but
cover the other interesting cases that this function must deal with,
either directly (with its own code) or indirectly (by calling cty
operation functions that handle it automatically).