cty: .Mark on an already-marked value shouldn't create nested marks
Previously we were correctly pulling all of the inner marks up to the top
level but _not_ discarding the existing "marker", and so we ended up with
a marker with another marker nested inside it.

That's confusing and violates assumptions elsewhere, so now we'll also
discard the existing marker and retain only what it was wrapping, so that
there can only ever be one level of marker wrapper.