function/stdlib: ZipmapFunc precise mark handling
ZipmapFunc can now handle marks on an element-by-element basis, to the
extent that the cty type system is able to do so.

Since map keys can't be individually sensitive, any sensitive values in
the keys list will aggregate on the map as a whole, but the individual
value elements will have their marks preserved individually.

We previously didn't have unit tests for ZipmapFunc so this also includes
some additional tests not related to marks, in order to cover the other
interesting cases this function must handle.