function/stdlib: MergeFunc precise handling of marks
MergeFunc will now only mark the resulting map or object with marks from
the given maps or objects as a whole, and not aggregate individual
element/attribute marks at the top-level result.

The individual element/attribute marks will still be preserved, but
they'll remain attached to the individual values they came from,
avoiding the result as a whole becoming marked.