function/stdlib: FlattenFunc precise handling of marks
FlattenFunc will now propagate marks from lists/sets/tuples which are
flattened to the top level result. Marks on non-flattened elements are
retained and not assigned to the resulting collection.

There is a significant change to the flattener helper function to ensure
that marks are propagated from all nested lists/sets/tuples, even after
first encountering an unknown collection. Previously finding an unknown
collection allowed us to terminate early, but to preserve marks we need
to continue to iterate through the rest of the arguments.