function/stdlib: SetproductFunc precise mark handling
Setproduct will now preserve marks on both the arguments and their
elements. Marks that apply to any entire argument will propagate to all
return values, while marks applied to a single element will remain
attached to individual values.

In the cases where this function returns a set result the marks still end
up aggregated at the root set value anyway due to cty's treatment of marks
in sets, but we can retain the precision in the cases where we'll return
a list.