function/stdlib: Fix setproduct bug with unknowns
If an argument to setproduct is a set with unknown values, its length is
unknown, as the unknown values may be duplicates of others in the set.
This should result in returning an unknown value propagating all marks.

Previously this crashed due to a misuse of `.Mark(marks)` instead of
`.WithMarks(marks)`. This commit also fixes an issue with the previous
logic, which would shortcut return and omit marks from any arguments
after the set with unknown length.

Also includes a couple of tests for setproduct that were in my local
working copy for some reason.