flatten: null dynamic values are known
When flattening a list of arguments, the special case of a `null`
dynamic value should not force the result to be unknown. We do not check
directly for "IsKnown" here, because it is specifically the dynamic type
which will cause `flatten` to become unknown, while unknown values alone
can be preserved preserved.