cty: Define our number range more precisely
Our documentation and implementation were previously pretty loose about
what _exactly_ the intended range of cty.Number is. Although internally
we store it as a *big.Float, most practical uses involve decimal string
representations of numbers, such as JSON serialization.

We'll now be fully explicit that we only expect to preserve precision for
the subset of numbers that are within the range of _both_ our in-memory
representation and of decimal string approximations of those values. This
was a practical constraint for many cases before anyway, but is now a
real documented constraint.

Along with this, the definition of "Equals" for numbers now honors the
definition by treating two numbers as equal only if they would have an
equal JSON serialization. cty can in principle be serialized in other
formats, including msgpack built-in, but JSON is the canonical
serialization in recognition of the fact that cty is typically used to
represent values destined for JSON-based network APIs.