cty: Allow capsule types inside set values
Previously cty would just panic if a caller tried to put values inside a
set whose element type is or contains a capsule type, because the set
rules lacked a hashing algorithm for those.

Now we have a hash implementation for capsule types. Ideally a capsule
type intended for use in sets should have a CapsuleOps that defines both
Equals and HashKey for the set implementation to rely on, but there is
also a less-efficient fallback behavior for types without HashKey (just
bucket them all together and scan) and a less-useful fallback behavior
for types without Equals (pointer equality of the encapsulated pointer,
as usual).

This now makes it safe to use capsule types in a system that also makes
use of sets of arbitrary types.