cty: sanity check signs of raw number before comparison
This fixes a performance issue when comparing large numbers to zero (or
large positive numbers to large negative numbers). Without this check,
the numbers are both converted to text, and string compared, which is
costly if the numbers are large.

This patch provides a cheap early-exit option if the signs don't match,
speeding up the cost of comparison to zero. As part of this, the .Modulo
function calling .RawEquals should also have significantly improved
performance (which was otherwise an expensive check for large numbers).