fix typo causing a panic in tuple unification
The `unifyObjectTypesToMap` function was inadvertently called from the
tuple-list unification function, which would panic since the value has
no attributes.

The given test example is possible speculatively unify, however we don't
have a codepath to recursively unify these yet, so we will just fail to
unify for now. This will fix the crashes in downstream consumers,
allowing the possibility of working around the shortcoming by using more
specific or different combinations of types.