convert: Only consider present map elements when converting to object type
Optional object attributes give us a new situation to consider here: the
source element type might not be compatible with all of the optional
attributes of a target object type, but that doesn't matter if the given
map doesn't include an element corresponding with the mismatching
attributes.

This is a bit awkward because we need to first allow the type conversion
logic to produce a valid conversion but then catch the attribute mismatch
only when applying the conversion function to the value. This is
effectively the same sequence of events that happens when converting from
string to number or string to bool: we optimistally assume that a
conversion will succeed when looking only at types, and then catch the
error dynamically once we have a final value to check.