convert: Test the exact error message strings
Producing intuitive error messages is part of the contract of this package,
so we should test exactly the error messages we're returning rather than
just whether conversion succeeded or not.

This exposed some cases where we _aren't_ producing good error messages
today, which are for the moment marked as FIXME because the purpose of
this change is just to improve the coverage of the test and not to change
any observable behavior.