walk: A more general value-transforming function
So far we've not had much success using the cty.Transform and
cty.TransformWithTransformer functions for any practical purpose, because
it's essentially impossible for them to correctly make any transformations
that affect the cty.Type of the input value without causing upstream
panics when reconstructing collections.

This new package introduces walk.TransformValue that aims to support
various different use-cases of transforming a cty.Value into something
else (which may or may not be a cty.Value) using a recursive walk while
giving the caller the ability to control the rules for merging various
nested values into a containing value.

When the return type of the transform is cty.Value this in particular
gives the caller an opportunity to try type unification when combining
many transformed nested values into a collection, whereas cty.Transform
and cty.TransformWithTransformer can only panic in that situation.

When the return type is something non-cty-related the walk still follows
the structure of the input value but the caller can decide how to combine
downstream values into upstream values as the function unwinds the stack.