cty: "Known-ness" represented internally by a Go type, not a Go value
Previously the internal definition of an unknown value was any value whose
raw value "v" field was exactly equal to the unexported global "unknown".

In preparation for supporting optional refinement of unknown values we'll
now broaden that definition to any value where "v" is an instance of
*unknownType, regardless of exact value.

In later commits we'll add optional fields inside struct unknownType to
capture the optional additional refinements. "Refinements" means extra
constraints on an unknown value that don't affect its type but do
constrain the possible concrete values the unknown value could be a
placeholder for.