cty: "Refinements" for unknown values
This is a new concept allowing for an unknown value to optionally carry
some additional information about constraints on the value for which it
is a placeholder.

Refinements do not affect the type of the value or most equality testing
rules. As of this commit they also don't affect any other operations on
values, but in future commits we'll use refinements to improve the
precision of certain operations on unknown values if the refinements give
enough information to still produce a known result.

For the moment this is limited just to a few key cases that seem likely
to help produce known results in common operations involving unknown
values:
- Equality tests will null, which are commonly used as part of conditional
  branches in calling applications.
- Numeric bounds, in particular focused on comparison to zero but general
  enough to help with other numeric comparisons too.
- Bounds on collection length, to use in conjunction with the numeric
  bounds so that asking for the length of an unknown collection can return
  an unknown number with numeric bounds.
- String prefixes, to help with string-prefix-related tests that seem to be
  a common subset of validation checks in calling applications.

Applications will tend to need to do some work to refine their own values
in order to benefit fully from this, but we'll also plumb refinements into
some of the built-in value operations and stdlib functions over time so
that even applications not working directly with refinements can still get
some indirect benefit.