cty: Value.Equal result is never null, and comparisons to non-null
Using the "refinements" feature we can refine the range of any unknown
values from Equal to exclude cty.NullVal(cty.Bool), since we know that
equality tests can never return null.

We can also return a known False if the test is between a null value and
an unknown value which has been refined as non-null.