cty: Simplify the collection length refinement model
I initially just copied the refinement model for numeric range, but in
retrospect that was silly because we know that collection lengths are
always integers and so there's no need to model unknown values and
fractional values and whatever else.

Instead, we now just have an integer lower and upper bound, always known
but when unconstrained they are the smallest and largest values of int
respectively.