msgpack: Round-trip some unknown value refinements
The new concept of "refinements" for unknown values allows us to track
some additional constraints on the range of an unknown value.

Normally those constraints would get lost during serialization, but for
msgpack in particular we can use an extension code to represent an unknown
value that has refinements and then include a serialization of the
refinements inside its body.

This then allows preserving the refinements when communicating over wire
protocols that use msgpack as their base serialization. However, as of
this commit the unmarshal step isn't quite complete becausethe refinement
API for numeric range is incomplete. We'll address both the refinement API
and the msgpack package's use of it in a future commit.

This required upgrading to a newer major version of the msgpack library
because v4 did not have all of the functionality required to encode a raw
extension object's payload.