cty: RefinementBuilder numeric bounds
The API for declaring numeric bounds was previously rather awkward and
inconsistent with the API for inspecting them in the ValueRange type.

Now we use separate methods for defining the lower bound and upper bound,
allowing the inclusiveness of each one to be independent. We retain the
previous NumberRangeInclusive helper for setting both bounds at once, but
it's now just a thin wrapper around the other two.