cty: Ignore numeric range refinements that don't add information
Previously we considered it a usage error to assert a bound that was
outside of the existing refined range.

However, if we already know that a number value is less than 2 then it
isn't really an error to assert that it's also less than 4; the second
assertion just doesn't give us any new information and so we should
silently discard it and retain the original assertion.

It's still an error to make assertions that directly contradict existing
refinements, such as asserting that an unknown value isn't null and then
later asserting that it's null.