cty: Value.Length constrains range of unknown results
When we're being asked for the length of an unknown collection we can
always put at least _some_ bounds on the range of the result, because we
know that a collection can't have a negative length and it cannot have a
length that is bigger than the current machine's integer size because
otherwise we wouldn't be able to index into it.

For some collections we may have further refinements on the length of the
unknown collection, in which case we'll also transfer those to being
refinements on the range of the numeric result.