cty: Value.LessThan and Value.GreaterThan use number range refinements
This allows us to return a known result in more cases than before by
making deductions based on the known numeric range.

This first implementation is a little more conservative than it strictly
needs to be because it just treats all bounds as exclusive rather than
taking into account inclusive bounds. This makes the implementation a bit
simpler for this initial round though, and should never generate an
incorrect known result.