function/stdlib: Strlen considers string prefix refinements
When given an unknown value, the Strlen function will now return a refined
unknown value with an inclusive lower bound. This bound is always at least
zero because a string can't have a negative length, but if we also know a
prefix for the string then we can refine the result even more to take that
into account.