cty: Refine string prefixes more conservatively
Because cty strings are unicode texts following both the normalization and
text segmentation rules, we need to be cautious in how we model prefix
matching of strings in the refinement system.

To avoid problems caused by combining character sequences, we'll trim off
characters from the end of a prefix if we cannot guarantee that they won't
combine with something that follows. This then avoids us overpromising
what the prefix will be in situations where the last character might
change after appending the rest of the string.

For applications that fully control their values and can guarantee that
no combining will occur at the boundary, the new function StringPrefixFull
retains the previous behavior of StringPrefix. The idea here is to be
conservative by default but allow applications to be less conservative if
they are able to offer greater guarantees.