function/stdlib: JSON functions can use string refinements
When encoding a JSON string we will often be able to predict the first
character of the result based just on the argument type.

That wouldn't be worthwhile alone, but we can also use a known first
character of a JSON document to infer a specific result type even for an
unknown value, and to raise an error immediately if the known prefix
cannot possibly be a JSON value of any type.

Unfortunately due to how JSON interacts with the cty type system the only
case where type information truly carries through in a round-trip is
encoding an unknown string and then decoding that unknown result, but
JSON may be produced and consumed by components other than this pair of
functions and they may be able to reach more definitive conclusions based
on this additional information.