convert: Refine conversions from structural to collection
When we're converting from a structural type to a collection type the
source type gives us information to refine the range of the resulting
collection even if the input value is unknown.

Similarly when converting from one collection type to another we can
transfer the length range constraints because, aside from set element
coalescing, conversions cannot change the length of a collection.