function/stdlib: SetProduct refines the length of unknown results
We know that the maximum possible cardinality for a resulting set is the
product of the cardinalities of all of the input sets. We might not know
the exact cardinalities of all of the input sets either, but if we do
know an upper bound then we can use that to calculate the maximum
cardinality of the result instead.