cty: Refine the ranges of arithmetic results
If we are performing addition, subtraction, or multiplication on unknown
numbers with known numeric bounds then we can propagate bounds to the
result by performing interval arithmetic.

This is not as complete as it could be because of trying to share a single
implementation across all of the functions while still dealing with all
of their panic edge cases.