function/stdlib: SortFunc always preserves the length of its input
If the list argument is an unknown value or contains unknown values then
we can't possibly return a fully-known result, but we do at least know
that sorting will never change the number of elements and so we can refine
our unknown result using the range of the input value.

The refinements system automatically collapses an unknown list collection
whose upper and lower length bounds are equal into a known list where
all elements are unknown, so this automatically preserves the known-ness
of the input length in the case where we're given a known list with
unknown elements, without needing to handle that as a special case here.