docs: An explicit policy for what "backward compatible" means
So far I've just left this implied because mostly I was the one
responsible for upgrading this library in various applications that use it
(by virtue of an employment coincidence).

However, the (indirect) use of this library has broadened lately, so it's
worth being clear about what does not constitute a breaking change in a
minor release so that application developers can have some guidance on
what to do when they see a test failure after upgrading and so that
hopefully they can avoid depending on implementation details in their main
code, outside of unit tests.