function: Don't call function Impl if we didn't call Type
By default cty function calls "short circuit" -- skip calling the Type
function and just immediately return cty.DynamicPseudoType -- if any of
the arguments are cty.DynamicVal.

However, in that case we were previously only skipping the call to Type
but yet still expecting Impl to be able to run. That's incorrect because
Impl functions should be able to treat Type as a "guard" and be guaranteed
that Impl will never run if Type failed.

To fix that hole we'll now track whether we skipped calling Type, and if
so we'll also skip calling Impl and just immediately return an unknown
value. Individual functions can still opt out of this behavior by
declaring on or more of their parameters as AllowDynamicType: true, in
which case their own Type function will get to decide how to handle that
situation.