msgpack: truncate long string prefix refinements
Prior to this commit, serialising an unknown value with a string prefix refinement with a sufficiently long prefix would result in an error on deserialisation, since cty limits the size of the refinements blob to 1024 bytes.

We choose an arbitrary maximum length of 256 bytes and truncate any
string prefix refinement on serialisation.

This behaviour is an exception to the general case where refined values
will round-trip losslessly.