cty: Value.Range should ignore marks
This fixes a panic when calling Range with a marked unknown string. Since the string is marked, it fails the unknownType type assertion on line 29, and then the call to AsString() on line 57 panics because AsString assumes the value is known.

The simplest fix for this is to simply unmark the value at the beginning of the function, since ValueRange makes no use of marks, and a caller making use of both value ranges and marks will retain the original value for any processing of marks.