cty: Explicit panic when using Value.Range with marked value
As usual with marks, we must panic whenever an operation might cause marks
to be silently lost. In such cases, it's always the calling application's
responsibility to unmark the value first and then ensure that it applies
the appropriate derived marks to whatever new value it's ultimately
returning based on the decision.

This was already a panic before, but now it's a panic with an explicit
message pointing directly at the problem.