don't use compact floats in msgpack
There is a bug in the compact floats implementation, which could cause
some integer values stored in a float to encode as an incorrect value.
The result of the expression used to check for compact integer encoding
is `float64(int64(n)) == n`, which is undefined when conversion is
not exact.