function: include marks when returning early with an unknown value
If a function parameter does not AllowUnknown, the function Call method
returns early with the appropriate unknown value. However, it was missing
any possible marks that were known from the given arguments.

We'll delay the early return slightly, so that all argument marks can be
collected and applied to the unknown return value.