function/stdlib: FormatFunc and FormatListFunc can handle DynamicPseudoType
We didn't previously include AllowDynamicType: true on the variadic
parameters for these functions and so the function system assumed that
this code couldn't deal with unspecified types at all.

However, the formatter is actually written to tolerate all possible values
and do _something_ reasonable with all of them, so we can safely opt in to
handle dynamic-typed arguments ourselves.

In particular that means that `cty.NullVal(cty.DynamicPseudoType)` is now
treated the same way as null values of known types, and `cty.DynamicVal`
is treated the same way as unknown values of known types.