cty: Fix various Value.HasElement quirks, and add tests
This function wasn't correctly handling some of the cases involving unknown
values, and it also apparently had no tests at all.

It now behaves consistently with our treatment of sets containing unknown
values elsewhere in cty, where we consider that to mean that the set could
contain any other value in addition to the ones that are known.