Forward to v1.4.1 (#7)
* Update CHANGELOG.md

* set: Add SameRules method to set.Rules

Previously, we checked if two sets had the same rules using a simple
equality check. This panics if the sets contain object types, as they
cannot be compared using `==`.

Adding a SameRules method to the Rules interface allows us to delegate
to the Type.Equals method where necessary, fixing this problem.

* stdlib: Fix set function crashes with empty sets

If one or more arguments to the stdlib set functions was an empty set of
dynamic pseudo type, the functions would panic due to incompatible set
rules.

We can special case empty dynamic pseudo type sets to be ignored through
type unification, because they are always capable of being converted to
any other type.

* Update CHANGELOG.md

* v1.4.1 release

* Update CHANGELOG.md for removal of json.UnmarshalDynamicWithImpliedType

---------

Co-authored-by: Martin Atkins <mart@degeneration.co.uk>
Co-authored-by: Alisdair McDiarmid <alisdair@users.noreply.github.com>